# nixos: doc: implement related packages in the manual (again)