###### What is this?

#32424 reverted by #33006 brought back to you with laziness and `meta.evaluates` of #33057.
Closes #33007.

###### Things done

- [X] Seems to work.
- [X] `configuration.nix(5)` manual builds on x86_64-linux.
- [ ] `configuration.nix(5)` manual builds on aarch64-linux.

Please check the last item with OfBorg before merging, I'm too lazy to rebuild aarch64 stdenv just to test this.

---

/cc @grahamc @vcunat