@oxij 2018-01-19
Well, it seems to work on my hacky cross-compiled aarch64 stdenv.

Ping @grahamc again because you was involved in 3 out of 4 related PRs.