@grahamc 2018-01-19
Will `relatedPackages` cover cases like `python3Packages.requests`? What will happen if a relatedPackage is deleted, and the reference is left dangling?