@oxij 2018-01-19
> One of the intentions behind making the inherits explicit here is so we don't need to make all the lib functions available in the top level, but instead say `lib.lists.compareLists`. If this seems like a reasonable thing to do, would you like to not `inherit` them here? It could be we're just not ready for that yet, seeing as I don't think many functions aren't exposed at the top level here.

Hm. I interpret those things differently. I though the intention of doing it this way was that you should inherit anything generic and not inherit things that are unlikely to be used out of the lib. Those new functions are pretty generic as proved by their usage out of the lib. 

Are you sure you want to (eventually) soil every nix module with a bunch of `with lib.<something>` at the top (like C++ does with `use <something>`)?
Why isn't `with lib` enough? I would prefer `lib` to be like Haskell's `Prelude` that gives the cohesive set of things you most likely want by default.

> Will `relatedPackages` cover cases like `python3Packages.requests`?

Not as a string because I didn't want to deal with parsing attributes (and the way `rename.nix` does it is super-ugly).
You can specify `{ name = "python3Packages.requests"; package = pkgs.python3Packages.requests; }`, though.

However, its is a valid question whenever something better can be invented, like, say, allowing both simple strings and ugly lists of `rename.nix`. I'll think about it.

> What will happen if a relatedPackage is deleted, and the reference is left dangling?

An error. I think that's a good thing. You are not supposed to delete packages that are referenced in documentation.

> Please provide a usage example for these functions.

The code I introduced next does provide it, but, okay, will fix.

> Also, is this really trivial?

Well, I know some people fear formulas, but it is the best way I have to describe what it does without giving in-words description of the actual implementation. Feel free to propose a better one.

> I'm worried about the `''` being confusing here, since it is identical to otherwise valid Nix syntax. Is there a better variable name you can imagine? 

Will fix.