@grahamc 2018-01-19
I agree with everything you said, sounds great! I look forward to the few remaining fixes :)