@oxij 2018-02-09
ping? Just as I said, its completely finished now.
I added a some more documentation beyond what was discussed and implemented the `x.y.z` with `rename.nix`-like lists. Seems perfect to me.