@oxij 2018-02-11
How do I profile this in nix?

How are you sure that the problem is with the whole sort and not with
recomputation of `hasPrefix` or with `compareLists`, for instance?
Normally I would imagine that sorting shouldn't introduce too much of a
problem.

I would guess the problem is that `optionListLess` is very inefficient
when evaluated as written. Can I ask nix to partially evaluate some
code? `cmp` can be partially evaluated into a much cheaper function.

I assume this evaluation time spike influences latency of channels, so
should I make an option to disable this sorting for now?

Which documentation nixos.org displays? From a latest succeeding master,
stable, something else? I would like to keep this sorting for all
publicly visible documentation, because its awesome.