@LnL7 2018-02-11
I used this and only checked the sort, it might be not be the sort itself but some other part.
Moving this sorting to the build should work, that way it will be sorted nicely for both the manual and manpages. But adding a flag that's only enabled for the manual also works as a workaround.

```
$ nix-build pkgs/top-level/release.nix -A metrics
$ grep .lapp result/nix-support/hydra-metrics
nixos.lapp.time 1.21916 s
nixos.lapp.maxresident 455256 KiB
nixos.lapp.allocations 273427272 B
nixos.lapp.values 4435159
```