@oxij 2018-02-11
> Moving this sorting to the build should work

Please elaborate, I don't understand what do you mean by that.