@vcunat 2018-02-11
https://nixos.org/nixos/manual/ shows the latest stable channel.