@vcunat 2018-02-11
> This significantly increased evaluation time

To be more explicit, it's basically a channel blocker for nixos-unstable ATM, as Hydra almost always runs out of memory during evaluation.  Unless we can find a solution very fast, I'd temporarily revert the problematic part until we find a better way.