@oxij 2018-02-11
> To be more explicit, it's basically a channel blocker for nixos-unstable ATM

Ok, this now has highest priority. Will either fix or add an option in the next 12 hours.