@oxij 2018-02-11
Ok, so.

On master (ad78e52357eb3df0f457c40185504567c7da524e) with offending patches applied:

nixos.smallContainer.time 7.811 s
nixos.smallContainer.maxresident 520740 KiB
nixos.smallContainer.allocations 1008637960 B
nixos.smallContainer.values 21474162
nixos.kde.time 12.343 s
nixos.kde.maxresident 527164 KiB
nixos.kde.allocations 1176678544 B
nixos.kde.values 23892286
nixos.lapp.time 8.115 s
nixos.lapp.maxresident 521496 KiB
nixos.lapp.allocations 1019577072 B
nixos.lapp.values 21652224

With a7d75ab6489dc5834e8402db374b4d0f1e774d53 reverted (which is untrivial):

nixos.smallContainer.time 2.34 s
nixos.smallContainer.maxresident 440616 KiB
nixos.smallContainer.allocations 262088816 B
nixos.smallContainer.values 4250508
nixos.kde.time 4.52 s
nixos.kde.maxresident 525820 KiB
nixos.kde.allocations 430129400 B
nixos.kde.values 6668632
nixos.lapp.time 2.372 s
nixos.lapp.maxresident 456340 KiB
nixos.lapp.allocations 273027928 B
nixos.lapp.values 4428570

I get almost exactly the same results by replacing the comparison function with `a: b: false`.
So the problem is not with sorting, but with inefficient comparisons.

With some trivial patches of `os/doc/efficient-comparisons` branch (https://github.com/oxij/nixpkgs/commits/os/doc/efficient-comparisons) I get:

nixos.smallContainer.time 3.294 s
nixos.smallContainer.maxresident 517396 KiB
nixos.smallContainer.allocations 365512360 B
nixos.smallContainer.values 6548134
nixos.kde.time 5.367 s
nixos.kde.maxresident 530352 KiB
nixos.kde.allocations 533552944 B
nixos.kde.values 8966258
nixos.lapp.time 3.178 s
nixos.lapp.maxresident 518004 KiB
nixos.lapp.allocations 376451472 B
nixos.lapp.values 6726196

The patches are pretty trivial and the comparison function stays the same.

So, in the light of that, the question.

a7d75ab6489dc5834e8402db374b4d0f1e774d53 costs ~3.5-4x in time and mallocs. This is, indeed, too much.
`os/doc/efficient-comparisons` now costs ~1.3-1.5x in time and mallocs without losing any code quality.
Is it still too much?

Should I butcher the understendability of `optionListLess` function in favor of more optimizations?
In theory, the sky is the limit here, since `builtins.sort` doesn't influence the evaluation much.

(Also, please don't merge `os/doc/efficient-comparisons` without advance notice, its WIP.)